Moved Program to AST.h
Program is generic and re-usable enough to be part of the root AST library.
Users should subclass it when using as in BrainFck. This gives them to oportunity to add extra members should they feel the need (e.g. regions for function/class definitions should a language become that complicated, looking at you Ork)