Improved logging performance
So turns out spdlog is not as speedy as I would have liked, after some tinkering, increasing the thread pool from 2 to 4 with a larger queue size improved this, for application performance.

However spdlog ignores the flush level for flushing to stdout  and stderr (this is dumb). This results in the stdout lagging behind the program MASSIVELY when the program executes quickly (e.g. for running a brainfck program).

This would be easily solved if only it respected the flush level (or at least gave to option to)